* contrib/tool/mlcc/Makefile.in: SF Patch #3530235 is merged. * doc/ja/README.ja, man/mlterm.1: Updated. * mlclient/main.c:   "--keepalive" is added to na_options. * ml_pty.h, ml_pty_ssh.c, main_loop.c, MLTermPty.c:   "ssh_keepalive_interval" option is added. * kik_mem.c: Minor fixes. * ml_pty.c: Declaration of ptsname() is enabled by defining _GNU_SOURCE.

+++ kiklib/src/kik_mem.c
@@ -512,7 +512,7 @@ kik_mem_dump_all(void)
 		{
 			fprintf( stderr , "%p(size %d , alloced at %s[l.%d in %s] is allocated.\n" ,
 				kik_iterator_indirect( iterator)->ptr ,
-				kik_iterator_indirect( iterator)->size ,
+				(int)kik_iterator_indirect( iterator)->size ,
 				kik_iterator_indirect( iterator)->func ,
 				kik_iterator_indirect( iterator)->line ,
 				kik_iterator_indirect( iterator)->file) ;
@@ -534,7 +534,7 @@ kik_mem_free_all(void)
 		{
 			fprintf( stderr , "%p(size %d , alloced at %s[l.%d in %s] is not freed.\n" ,
 				kik_iterator_indirect( iterator)->ptr ,
-				kik_iterator_indirect( iterator)->size ,
+				(int)kik_iterator_indirect( iterator)->size ,
 				kik_iterator_indirect( iterator)->func ,
 				kik_iterator_indirect( iterator)->line ,
 				kik_iterator_indirect( iterator)->file) ;
